# Eaas/openapi.yaml
openapi: 3.0.3
info:
  title: EaaS — Escrow as a Service API
  description: |
    API-first, Africa-native escrow platform. Enables secure transactions
    between buyers and sellers by holding funds in escrow until delivery is confirmed.
    
    ## Authentication
    - **Customer/Admin:** Bearer JWT token via `Authorization: Bearer <token>`
    - **Merchant API:** API Key via `X-API-Key: sk_live_xxxx` header
    - **Webhooks from PSP:** Verified via `X-Interswitch-Signature` HMAC header
    
    ## Idempotency
    All POST requests that create resources must include `X-Idempotency-Key` header.
    
    ## Rate Limiting
    - 100 requests/minute per API key
    - Rate limit headers returned: `X-RateLimit-Limit`, `X-RateLimit-Remaining`
    
  version: 1.0.0
  contact:
    name: EaaS Developer Support
    email: developers@eaas.africa
servers:
  - url: https://api.eaas.africa/api/v1
    description: Production
  - url: https://sandbox.eaas.africa/api/v1
    description: Sandbox
...
paths:
...
  /escrow:
    post:
      tags: [Escrow]
      summary: Create a new escrow transaction
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - in: header
          name: X-Idempotency-Key
          required: true
          schema:
            type: string
          description: Unique key to prevent duplicate transactions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEscrowRequest'
            example:
              merchantId: "merchant-uuid"
              productDescription: "Samsung Galaxy S24 Ultra 256GB — Phantom Black"
              productQuantity: 1
              amount: 850000.00
              currency: "NGN"
              agreedDeliveryDays: 5
      responses:
        '201':
          description: Escrow created
          content:
            application/json:
              example:
                status: "success"
                data:
                  reference: "TXN-2024-A3F9K2"
                  amount: 850000.00
                  escrowFee: 12750.00
                  merchantAmount: 837250.00
                  currency: "NGN"
                  paymentLink: "https://paystack.com/pay/TXN-2024-A3F9K2"
                  status: "INITIATED"
                  expiresAt: "2024-01-12T12:00:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    get:
      tags: [Escrow]
      summary: List my escrow transactions (paginated)
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [INITIATED, FUNDED, SHIPPED, DELIVERED, CONFIRMED, DISPUTED, RELEASED, REFUNDED]
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 20
        - in: query
          name: fromDate
          schema:
            type: string
            format: date
        - in: query
          name: toDate
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Paginated list of escrow transactions
          content:
            application/json:
              example:
                status: "success"
                data:
                  content: []
                  totalElements: 150
                  totalPages: 8
                  page: 0
                  size: 20
  /escrow/{reference}:
    get:
      tags: [Escrow]
      summary: Get escrow transaction by reference
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - in: path
          name: reference
          required: true
          schema:
            type: string
          example: "TXN-2024-A3F9K2"
      responses:
        '200':
          description: Escrow transaction details
        '404':
          $ref: '#/components/responses/NotFound'
...
  /escrow/{reference}/ship:
    post:
      tags: [Escrow]
      summary: Merchant submits tracking number (marks as SHIPPED)
      security:
        - apiKeyAuth: []
        - bearerAuth: []
      parameters:
        - in: path
          name: reference
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [trackingNumber, logisticsProvider]
              properties:
                trackingNumber:
                  type: string
                  example: "GIG-2024-XY9182"
                logisticsProvider:
                  type: string
                  example: "GIG Logistics"
                estimatedDeliveryDate:
                  type: string
                  format: date
                  example: "2024-01-14"
      responses:
        '200':
          description: Escrow marked as SHIPPED
          content:
            application/json:
              example:
                status: "success"
                data:
                  reference: "TXN-2024-A3F9K2"
                  status: "SHIPPED"
                  trackingNumber: "GIG-2024-XY9182"
                  shippedAt: "2024-01-12T09:30:00Z"
        '409':
          $ref: '#/components/responses/Conflict'
...
  /escrow/{reference}/confirm:
    post:
      tags: [Escrow]
      summary: Customer confirms receipt — triggers payout to merchant
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: reference
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Delivery confirmed — payout triggered
          content:
            application/json:
              example:
                status: "success"
                message: "Delivery confirmed. Merchant will receive payment within 24 hours."
                data:
                  reference: "TXN-2024-A3F9K2"
                  status: "CONFIRMED"
                  confirmedAt: "2024-01-15T10:00:00Z"
...
  /escrow/{reference}/cancel:
    post:
      tags: [Customer Escrow Operations]
      summary: Customer cancels escrow
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: reference
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Escrow cancelled
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Cannot cancel — escrow already funded or beyond

  # ─── DISPUTES ──────────────────────────────────────────────
...
  /admin/kyc/{merchantId}/approve:
    put:
      tags: [Admin]
      summary: Approve merchant KYC
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: merchantId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: KYC approved — API key generated and emailed
  /admin/kyc/{merchantId}/reject:
    put:
      tags: [Admin]
      summary: Reject merchant KYC
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: merchantId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: KYC rejected — merchant notified
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          example:
            status: "error"
            message: "Validation failed"
            errors:
              - field: "amount"
                message: "Amount must be greater than zero"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          example:
            status: "error"
            message: "Merchant not found"
    Conflict:
      description: State conflict
      content:
        application/json:
          example:
            status: "error"
            message: "Cannot transition from CONFIRMED to DISPUTED"
  schemas:
...
    CreateEscrowRequest:
      type: object
      required: [merchantId, productDescription, amount]
      properties:
        merchantId:
          type: string
          format: uuid
        productDescription:
          type: string
          maxLength: 1000
        productQuantity:
          type: integer
          minimum: 1
          default: 1
        amount:
          type: number
          minimum: 100
          description: Minimum escrow amount is ₦100
        agreedDeliveryDays:
          type: integer
          minimum: 1
          maximum: 30
          default: 7
...